Refactor(frontend): Moved views into directories
Views were previously in singular files which got large especially for
the `Culture` view. By moving towards directories we can more easily put
associated components, tests, etc with their respected views overall
leading to better organization.
