+++ frontend/views/Culture/ToolsFAB.tsx
@@ -1,0 +1,45 @@
+import React, { useState } from "react";
+import { StyleSheet } from "react-native";
+
+import { FAB } from "react-native-paper";
+
+const Styles = StyleSheet.create({
+  fab: {
+    // TODO: Emulate Fixed for the Floating Action Button
+  },
+});
+
+/**
+ * Properties for {@link ToolsFAB}
+ */
+type ToolsFABProps = {
+  // onSave function called when hitting the "save" button. This button should
+  // toggle the whether this FAB is being displayed.
+  onSave: () => void;
+  // onAdd function called when hitting the "add" button.
+  onAdd: () => void;
+};
+
+/**
+ * ToolsFAB displays a {@link FAB.Group} that has two sub {@link FAB} one for editing and one for saving.
+ *
+ * @param {ToolsFABProps} props
+ * @returns {React.ReactElement} React component
+ */
+export default function ToolsFAB(props: ToolsFABProps): React.ReactElement {
+  const [open, setOpen] = useState(false);
+
+  return (
+    <FAB.Group
+      visible={true}
+      style={Styles.fab}
+      open={open}
+      icon={open ? "close" : "wrench"}
+      actions={[
+        { icon: "plus", onPress: () => props.onAdd() },
+        { icon: "content-save", onPress: () => props.onSave() },
+      ]}
+      onStateChange={() => setOpen(!open)}
+    />
+  );
+}
