+++ frontend/views/Culture/index.ts
@@ -1,0 +1,3 @@
+import Culture from "./Culture";
+
+export default Culture;
