+++ frontend/views/EditInsight/index.ts
@@ -1,0 +1,3 @@
+import EditInsight from "./EditInsight";
+
+export default EditInsight;
