+++ frontend/views/Home/index.ts
@@ -1,0 +1,3 @@
+import Home from "./Home";
+
+export default Home;
