+++ frontend/views/Recovery/index.ts
@@ -1,0 +1,3 @@
+import Recovery from "./Recovery";
+
+export default Recovery;
