+++ frontend/views/Register/Register.tsx
@@ -15,14 +15,11 @@ import { CommonActions } from "@react-navigation/native";
 import { RouteProp } from "@react-navigation/native";
 import { useFormik } from "formik";
 
-import { Admin, AuthPayload } from "../api";
-import { updateUser, Store } from "../redux";
-import { Routes } from "../routes";
-import {
-  RegisterHelpText,
-  RegisterValidationSchema,
-  TermsOfServiceURL,
-} from "../constants";
+import { Admin, AuthPayload } from "../../lib";
+import { updateUser, Store } from "../../redux";
+import { Routes } from "../../routes";
+import { RegisterHelpText, TermsOfServiceURL } from "../../constants";
+import Validation from "./validation";
 
 type Props = {
   navigation: StackNavigationProp<Routes, "Register">;
@@ -94,7 +91,7 @@ function Register(props: Props): React.ReactElement {
     touched,
     handleSubmit,
   } = useFormik({
-    validationSchema: RegisterValidationSchema,
+    validationSchema: Validation,
     initialValues: initialValues,
     onSubmit: (values) => register(values),
   });
