+++ frontend/views/Register/index.ts
@@ -1,0 +1,3 @@
+import Register from "./Register";
+
+export default Register;
